fix review, move it into 'component' folder
