+++ libs/event/src/lib/components/explanation/explanation.component.ts
