+++ libs/event/src/lib/components/explanation/explanation.module.ts
